LayoutTests/platform/ios-sim/fast/files/file-reader-sandbox-iframe-expected.txt

CONSOLE MESSAGE: line 9: Received data: 0123456789
Test that FileReader works under the sandboxed iframe.
